* main: (227 commits)
  💚 CI 빌드 수정 : 메인 브랜치 푸쉬 이벤트 발생 시 CI 진행이 되지 않는 문제 해결 - 메인 브랜치 푸쉬 이벤트 발생 시 CI 진행이 되지 않는 버그 수정
  💚 CI 빌드 수정 : 워크플로우 Name명 수정 - 가독성을 위해 명확하게 수정 작업.
  💚 CI 빌드 수정 : PR 생성 시 CI 자동 실행 버그 해결에 관한 설정 내용 추가 수정
  💚 CI 빌드 수정 : PR 생성 시 CI 자동 실행 버그 해결
  🐛	버그 수정 : 기술퀴즈 문제풀고 정답 맞췄을때 userQuizStatus가 solved로 변하지 않는 버그 수정
  🐛	버그 수정 : 사용자의 프로핖이 solved.ac에 등록되어 있지 않을 경우 사용자 프로필을 null로 받아오게되어 생기는 버그 수정
  ✨ 새 기능 : Prometheus와 Grafana 모니터링 설정을 위한 초기 구성 - build.gradle: `spring-boot-starter-actuator`와 `micrometer-registry-prometheus` 의존성 추가 - application.yml: Prometheus 메트릭 수집을 위한 Actuator 엔드포인트 활성화 - Spring Security: `/actuator/**` 경로 접근 권한 추가
  ☔️테스트 : 이력서 create 코드 추가
  👷 CI 빌드 수정 : GitHub Action Docker Cache 최적화
  👷 CI 빌드 수정 : GitHub Action CICD Discord 알림 기능 성공 확인 후 로직 정상화
  👷 CI 빌드 수정 : GitHub Action CICD Discord 실패 알림 기능 수정
  👷 CI 빌드 수정 : GitHub Action CICD Discord 알림 기능 테스트 및 yml 구분을 위해 PR 알림봇 yml 이름 수정 - 헤깔리지 않기 위해 알림봇 이름 수정.
  👷 CI 빌드 추가 : GitHub Action CICD Discord 알림 기능 추가
  👷 CI 빌드 추가 : GitHub Action EC2 SSH Deploy 추가
  👷 CI 빌드 추가 : GitHub Action Docker Push 수정
  👷 CI 빌드 추가 : GitHub Action Docker Push 추가
  CI 빌드 추가 : GitHub Action 빌드 테스트
  👷 CI 빌드 추가 : GitHub Action 빌드 테스트
  👷 CI 빌드 수정 : GitHub Action 빌드 테스트
  👷 CI 빌드 추가 : GitHub Action 빌드 테스트
  ...

+++ src/main/java/darkoverload/itzip/feature/algorithm/domain/Problem.java
@@ -15,7 +15,7 @@ public class Problem {
     private Long problemId;
     private String title;
     private Integer level;
-    private Integer acceptedUserCount;
+    private Long acceptedUserCount;
     private Integer averageTries;
 
     public ProblemEntity convertToEntity() {
